requirements.txt
Pillow==9.0.0
pydirectinput==1.0.4
pygetwindow==0.0.9
pytesseract==0.3.10
